Reset to commit 74f77f4 state
