--- src/main/java/ch/uzh/ifi/hase/soprafs24/entity/Stock.java
@@ -1,74 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.entity;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "STOCK")
-public class Stock {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String symbol; // e.g. specific stock name 
-
-    @Column(nullable = false)
-    private Double price;
-
-    @Column(nullable = false)
-    private LocalDate date;
-
-    @Column
-    private String currency;
-
-    @Column
-    private Long volume;
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getSymbol() {
-        return symbol;
-    }
-
-    public void setSymbol(String symbol) {
-        this.symbol = symbol;
-    }
-
-    public Double getPrice() {
-        return price;
-    }
-
-    public void setPrice(Double price) {
-        this.price = price;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public void setDate(LocalDate date) {
-        this.date = date;
-    }
-
-
-    public String getCurrency() {
-        return currency;
-    }
-
-    public void setCurrency(String currency) {
-        this.currency = currency;
-    }
-    public Long getVolume() {
-        return volume;
-    }
-
-    public void setVolume(Long volume) {
-        this.volume = volume;
-    }
-    
-}
