--- src/main/java/ch/uzh/ifi/hase/soprafs24/repository/GameRepository.java
@@ -1,11 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.repository;
-
-import ch.uzh.ifi.hase.soprafs24.entity.Game;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface GameRepository extends JpaRepository<Game, Long> {
-   // Optional<Game> findByLobbyId(Long lobbyId);
-   // Game findByGameID(Long gameID);
-}
