--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/GameGetDTO.java
@@ -1,15 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-
-public class GameGetDTO {
-
-    private Long gameID;
-
-    public Long getGameID() {
-        return gameID;
-    }
-
-    public void setGameID(Long gameID) {
-        this.gameID = gameID;
-    }
-    
-}
