--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/GamePostDTO.java
@@ -1,15 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-
-public class GamePostDTO {
-
-
-    private Long lobbyId;
-
-    public Long getLobbyId() {
-        return lobbyId;
-    }
-
-    public void setLobbyId(Long lobbyId) {
-        this.lobbyId = lobbyId;
-    }
-}
