+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/LobbyUserPostDTO.java
@@ -1,0 +1,15 @@
+package ch.uzh.ifi.hase.soprafs24.rest.dto;
+
+public class LobbyUserPostDTO {
+    
+    private Long userId;
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+}
+
