--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/StockPriceGetDTO.java
@@ -1,32 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-
-public class StockPriceGetDTO {
-
-    private String symbol;
-    private int round;
-    private double price;
-
-    public String getSymbol() {
-        return symbol;
-    }
-
-    public void setSymbol(String symbol) {
-        this.symbol = symbol;
-    }
-
-    public int getRound() {
-        return round;
-    }
-
-    public void setRound(int round) {
-        this.round = round;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-}
