--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/TransactionRequestDTO.java
@@ -1,31 +1,0 @@
-//SJ: game package is centered on game logic (GameManager, PlayerState, Transaction)
-//DTOs should be data containers for HTTP communication 
-//This is not a class that does the actual game logic,transaction request that the client (front) sends to the backend. so i thought it is more right to put it in the dto folder. 
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-
-
-public class TransactionRequestDTO {
-    private String stockId;
-    private int quantity;
-    private String type; // BUY or SELL
-
-    public String getStockId() {
-        return stockId;
-    }
-    public void setStockId(String stockId) {
-        this.stockId = stockId;
-    }
-    public int getQuantity() {
-        return quantity;
-    }
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public String getType() {
-        return type;
-    }
-    public void setType(String type) {
-        this.type = type;
-    }
-}
