--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/UserLoginDTO.java
@@ -1,24 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-
-public class UserLoginDTO {
-
-  private String password;
-  private String username;
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
-  }
-
-  public String getPassword() {
-    return password;
-  }
-
-  public void setPassword(String password) {
-    this.password = password;
-  }
-
-}
